scripts/image_pipeline/requirements.txt
google-genai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
python-dotenv>=1.0.0
